+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 #### Enhancements
 * check for jwt token algorithms used to reject "none" algorithm 
 * update dependencies to fix vulnerabilities in elliptic
- 
+* add helper script to generated bcrypt password hash and allow setting http auth password hash from file inside docker, #434
+
 ## Version 0.7.1
 #### Bugfixes
 * update handling of big numbers displayed as json formatted values. For big numbers wrong values may be shown, #400